pyproject.toml
[project]
name = "qaguru-advanced"
version = "0.1.0"
description = "Домашние задания на курс https://qa.guru/python-advanced"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pytest>=8.4.1",
    "ruff>=0.12.10",
    "uvicorn>=0.35.0",
    "curlify2>=2.0.0",
    "allure-pytest>=2.15.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
    "fastapi-pagination>=0.13.3",
    "email-validator>=2.2.0",
    "mypy>=1.17.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
]
